type changed back to int

+++ lab_02_02_01/main.c
@@ -12,7 +12,7 @@
 // function receives pointer to an array and number of elements in this array
 // function returns 1, if input is correct
 // 0, if input in not correct
-int input(long long int *array, int num_of_elem)
+int input(int *array, int num_of_elem)
 {
     int input_is_correct = YES;
     for (int i = 0; i < num_of_elem; i++)
@@ -30,7 +30,7 @@ int input(long long int *array, int num_of_elem)
 // function receives int x and returns 1 if:
 // 1. x <= 1, 2. x is prime;
 // otherwise function returns 0
-int is_prime(long long int x)
+int is_prime(int x)
 {
     int x_is_prime = YES;
     
@@ -57,11 +57,11 @@ int main()
     if (corr_input != 1 || number_of_elements <= 0 || number_of_elements > ARRAY_SIZE)
         return INCORRECT_INPUT;
     
-    long long int arr[ARRAY_SIZE];
+    int arr[ARRAY_SIZE];
     if (input(arr, number_of_elements) == NO)
         return INCORRECT_INPUT;
     
-    long long int result_array[ARRAY_SIZE];
+    int result_array[ARRAY_SIZE];
     int count_new_array = 0;
     for (int i = 0; i < number_of_elements; i++)
     {
